+++ src/components/ComicDetailsPanel/ComicDetailsPanel.tsx
@@ -242,7 +242,6 @@ function ComicDetailsPanel({
                 <ErrorPresenter error={comicDataError} />
             )}
             <ItemNavigation
-                currentComic={currentComic}
                 itemNavigationData={comicItems ?? []}
                 isLoading={isLoadingInitial}
                 isFetching={isFetching}
@@ -324,7 +323,6 @@ function ComicDetailsPanel({
                     <>
                         <hr className="my-4 mx-0 border-solid border-b max-w-none" />
                         <FilteredNavigationData
-                            currentComic={currentComic}
                             isLoading={isLoadingInitial}
                             isFetching={isFetching}
                             isSaving={isSaving}