+++ src/components/ComicDetailsPanel/ItemNavigation/ItemNavigation.tsx
@@ -4,7 +4,6 @@ import InlineSpinner from '@components/InlineSpinner'
 import NavElement, { NavElementMode } from '@components/NavElement/NavElement'
 import Spinner from '@components/Spinner'
 import useItemNavigationDataByType from '@hooks/useItemNavigationDataByType'
-import { ComicId } from '@models/ComicId'
 import { HydratedItemNavigationData } from '@models/HydratedItemData'
 import { ItemId } from '@models/ItemId'
 
@@ -12,7 +11,6 @@ import { PickEnum } from '~/tsUtils'
 
 export default function ItemNavigation({
     itemNavigationData,
-    currentComic,
     isLoading,
     isFetching,
     useColors,
@@ -24,7 +22,6 @@ export default function ItemNavigation({
     onAddItem,
 }: {
     itemNavigationData: HydratedItemNavigationData[]
-    currentComic: ComicId
     isLoading: boolean
     isFetching: boolean
     useColors: boolean
@@ -46,7 +43,6 @@ export default function ItemNavigation({
             return (
                 <NavElement
                     key={item.id}
-                    currentComic={currentComic}
                     item={item}
                     onSetCurrentComic={onSetCurrentComic}
                     useColors={useColors}
@@ -66,7 +62,6 @@ export default function ItemNavigation({
             onSetCurrentComic,
             onShowInfoFor,
             useColors,
-            currentComic,
         ]
     )