+++ src/components/EditorModePanel/EditorModePanel.tsx
@@ -268,7 +268,6 @@ function EditorModePanel({
                 <>
                     <MissingNavElement
                         navigationData={editorData.missing.cast}
-                        currentComic={currentComic}
                         title="Missing cast"
                         description="Navigate to comics without cast members"
                         onSetCurrentComic={setCurrentComic}
@@ -277,7 +276,6 @@ function EditorModePanel({
                     />
                     <MissingNavElement
                         navigationData={editorData.missing.location}
-                        currentComic={currentComic}
                         title="Missing location"
                         description="Navigate to comics without locations"
                         onSetCurrentComic={setCurrentComic}
@@ -286,7 +284,6 @@ function EditorModePanel({
                     />
                     <MissingNavElement
                         navigationData={editorData.missing.storyline}
-                        currentComic={currentComic}
                         title="Missing storyline"
                         description="Navigate to comics without storylines"
                         onSetCurrentComic={setCurrentComic}
@@ -295,7 +292,6 @@ function EditorModePanel({
                     />
                     <MissingNavElement
                         navigationData={editorData.missing.title}
-                        currentComic={currentComic}
                         title="Missing title"
                         description="Navigate to comics without a title"
                         onSetCurrentComic={setCurrentComic}
@@ -304,7 +300,6 @@ function EditorModePanel({
                     />
                     <MissingNavElement
                         navigationData={editorData.missing.tagline}
-                        currentComic={currentComic}
                         title="Missing tagline"
                         description="Navigate to comics without a tagline"
                         onSetCurrentComic={setCurrentComic}