+++ src/components/FilteredNavigationData/FilteredNavigationData.tsx
@@ -11,7 +11,6 @@ import { ItemId } from '@models/ItemId'
 import { ItemType } from '@models/ItemType'
 
 export default function FilteredNavigationData({
-    currentComic,
     isLoading,
     isFetching,
     isSaving,
@@ -23,7 +22,6 @@ export default function FilteredNavigationData({
     editMode,
     onAddItem,
 }: {
-    currentComic: ComicId
     isLoading: boolean
     isFetching: boolean
     isSaving: boolean
@@ -73,7 +71,6 @@ export default function FilteredNavigationData({
             {(filteredItemData.length > 0 || isLoading) && !hasError ? (
                 <div className="overflow-y-auto overflow-x-hidden max-h-52">
                     <ItemNavigation
-                        currentComic={currentComic}
                         itemNavigationData={filteredItemData}
                         isLoading={isLoading}
                         isFetching={isFetching}