+++ src/components/NavElement/NavElement.tsx
@@ -1,9 +1,39 @@
+import { ConnectedProps, connect } from 'react-redux'
+
 import NavButton from '@components/ComicDetailsPanel/NavButton/NavButton'
-import { ComicId } from '@models/ComicId'
 import { HydratedItemNavigationData } from '@models/HydratedItemData'
 import { ItemId } from '@models/ItemId'
+import { skipToken } from '@reduxjs/toolkit/dist/query'
+import { useRandomComicQuery } from '@store/api/itemApiSlice'
+import { AppDispatch, RootState } from '@store/store'
 
 import { createTintOrShade } from '~/color'
+
+import { useAlternateLayout } from './useAlternateLayout'
+
+const mapState = (state: RootState) => {
+    return {
+        settings: state.settings.values,
+        currentComic: state.comic.current,
+    }
+}
+
+const mapDispatch = (_dispatch: AppDispatch) => {
+    return {}
+}
+
+const connector = connect(mapState, mapDispatch)
+type PropsFromRedux = ConnectedProps<typeof connector>
+type NavElementProps = PropsFromRedux & {
+    item: HydratedItemNavigationData
+    onSetCurrentComic: (_: number) => void
+    useColors: boolean
+    onShowInfoFor: (_: ItemId) => void
+    mode: NavElementMode
+    editMode?: boolean
+    onRemoveItem?: (_: ItemId) => void
+    onAddItem?: (_: ItemId) => void
+}
 
 export enum NavElementMode {
     Present,
@@ -12,33 +42,36 @@ export enum NavElementMode {
     Editor,
 }
 
-export default function NavElement({
+export function NavElement({
+    settings,
+    currentComic,
     item,
     useColors,
-    currentComic,
     onSetCurrentComic,
     onShowInfoFor,
     mode,
     editMode,
     onRemoveItem,
     onAddItem,
-}: {
-    item: HydratedItemNavigationData
-    currentComic: ComicId
-    onSetCurrentComic: (_: number) => void
-    useColors: boolean
-    onShowInfoFor: (_: ItemId) => void
-    mode: NavElementMode
-    editMode?: boolean
-    onRemoveItem?: (_: ItemId) => void
-    onAddItem?: (_: ItemId) => void
-}) {
+}: NavElementProps) {
     let backgroundColor = item.color
     if (!backgroundColor.startsWith('#')) {
         backgroundColor = `#${backgroundColor}`
     }
     const foregroundColor = createTintOrShade(item.color)
     const hoverFocusColor = createTintOrShade(item.color, 2)
+
+    const { data: randomComic, refetch: refreshRandomComic } =
+        useRandomComicQuery(
+            mode === NavElementMode.Present && settings?.showItemRandomButton
+                ? {
+                      currentComic,
+                      itemId: item.id,
+                      skipGuest: settings.skipGuest,
+                      skipNonCanon: settings.skipNonCanon,
+                  }
+                : skipToken
+        )
 
     let extraStuff = <></>
     if (editMode) {
@@ -84,13 +117,17 @@ export default function NavElement({
         }
     }
 
+    const [alternateLayout, selfRef] = useAlternateLayout()
+
     return (
         <>
             <div
+                ref={selfRef}
                 id={`qc-ext-navelement-${item.id}`}
                 className={
                     `qc-ext-navelement flex items-center rounded` +
-                    (useColors ? ' with-color' : '')
+                    (useColors ? ' with-color' : '') +
+                    (alternateLayout ? ' flex-wrap justify-between' : '')
                 }
                 style={
                     useColors
@@ -116,7 +153,10 @@ export default function NavElement({
                     onSetCurrentComic={onSetCurrentComic}
                 />
                 <button
-                    className="font-bold flex-auto py-1"
+                    className={
+                        'font-bold flex-auto py-1' +
+                        (alternateLayout ? ' -order-1 basis-full' : '')
+                    }
                     onClick={(e) => {
                         e.preventDefault()
                         onShowInfoFor(item.id)
@@ -142,7 +182,21 @@ export default function NavElement({
                     faClass="fast-forward"
                     onSetCurrentComic={onSetCurrentComic}
                 />
+                {mode === NavElementMode.Present &&
+                    settings?.showItemRandomButton && (
+                        <NavButton
+                            comicNo={randomComic ?? 0}
+                            title={`Random strip with ${item.shortName}`}
+                            faClass="question"
+                            onSetCurrentComic={(c) => {
+                                onSetCurrentComic(c)
+                                refreshRandomComic()
+                            }}
+                        />
+                    )}
             </div>
         </>
     )
 }
+
+export default connector(NavElement)