+++ src/components/SettingsDialog/SettingsPanel/SettingsPanel.tsx
@@ -77,6 +77,16 @@ export default function SettingsPanel({
             />
             <ToggleSetting
                 settings={settings}
+                setting="showItemRandomButton"
+                updateSettings={updateSettings}
+                label="Show button for going to random comic per item"
+                description={
+                    'Enabling this setting will add a button for navigation to a random comic containing the specific ' +
+                    'item.'
+                }
+            />
+            <ToggleSetting
+                settings={settings}
                 setting="showTaglineAsTooltip"
                 updateSettings={updateSettings}
                 label="Show tagline as comic tooltip"