+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Add Storybook link to README
 -   Added list of comics an item is featured in. Closes [#19](https://github.com/Questionable-Content-Extensions/client/issues/19)
 -   Added feature and setting for whether to show the current comic's tagline as its tooltip. Closes [#47](https://github.com/Questionable-Content-Extensions/client/issues/47)
+-   Added feature and setting for being able to navigate to random comics by item. Closes [#20](https://github.com/Questionable-Content-Extensions/client/issues/20)
 
 ### Changed 🔧